% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pontential_in_complete_graph_case}
\alias{pontential_in_complete_graph_case}
\title{First example}
\usage{
pontential_in_complete_graph_case(origin, dest, trips, unique_geozomes)
}
\arguments{
\item{args}{string vector}
}
\description{
First example
}
