% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pontential_on_graph}
\alias{pontential_on_graph}
\title{First example}
\usage{
pontential_on_graph(vertex1, vertex2, netflow_R, unique_geozomes)
}
\arguments{
\item{args}{string vector}
}
\description{
First example
}
