% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HodgePotentialHumanFlow-package.R
\name{scalar_potential_on_complete_graph}
\alias{scalar_potential_on_complete_graph}
\title{This function calculates scalar potential of human flow for a given origin-destination matrix (complete (graph version).}
\usage{
scalar_potential_on_complete_graph(od_table)
}
\arguments{
\item{od_table}{A data.frame with three columns: origin, dest, and trips}
}
\value{
a data.frame with two columns: zone and potential
}
\description{
This function calculates scalar potential of human flow for a given origin-destination matrix (complete (graph version).
}
