% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HodgePotentialHumanFlow-package.R
\name{scalar_potential_on_complete_graph_with_pvalue}
\alias{scalar_potential_on_complete_graph_with_pvalue}
\title{This function calculates scalar potential of human flow for a given origin-destination matrix with p-value (complete graph version).}
\usage{
scalar_potential_on_complete_graph_with_pvalue(od_table)
}
\arguments{
\item{od_table}{A data.frame with three columns: origin, dest, and trips}
}
\value{
This function returns a data.frame with three columns: zone, potential, p-value
}
\description{
This function calculates scalar potential of human flow for a given origin-destination matrix with p-value (complete graph version).
}
