% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HodgePotentialHumanFlow-package.R
\name{scalar_potential_on_graph}
\alias{scalar_potential_on_graph}
\title{This function calculates scalar potential of human flow for a given origin-destionation matrix (graph version)}
\usage{
scalar_potential_on_graph(flow_on_edges)
}
\arguments{
\item{flow_on_edges}{A data.frame with three columns: vertex1, vertex2, and netflow}
}
\value{
This function returns a list:
\itemize{
\item value: a data.frame with three columns: zone, potential, p-value
\item R2: percentage of gradient component
}
}
\description{
This function calculates scalar potential of human flow for a given origin-destionation matrix (graph version)
}
