% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HodgePotentialHumanFlow-package.R
\name{scalar_potential_on_weighted_graph}
\alias{scalar_potential_on_weighted_graph}
\title{This function calculates scalar potential of human flow for a given origin-destination matrix with p-value (weighted graph version)}
\usage{
scalar_potential_on_weighted_graph(
  flow_on_edges,
  num_samples = 1e+05,
  seed = -1
)
}
\arguments{
\item{flow_on_edges}{A data.frame with three columns: vertex1, vertex2, weight, and netflow.
\itemize{
\item num_samples: Monte calro samples (default = 1e5)
\item seed: Monte calro samples (default = -1, then generate it by std::random_device)
}}
}
\value{
This function returns a data.frame with three columns: zone, potential, p-value
}
\description{
This function calculates scalar potential of human flow for a given origin-destination matrix with p-value (weighted graph version)
}
