% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HodgePotentialHumanFlow-package.R
\name{scalar_potential_with_details}
\alias{scalar_potential_with_details}
\title{This function calculates scalar potential of human flow for a given origin-destionation matrix with details}
\usage{
scalar_potential_with_details(od_table)
}
\arguments{
\item{od_table}{A data.frame with three columns: origin, dest, and trips}
}
\value{
This function returns a list:
\itemize{
\item value: a data.frame with three columns: zone, potential, p-value
\item R2: percentage of gradient component
}
}
\description{
This function calculates scalar potential of human flow for a given origin-destionation matrix with details
}
