% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HodgePotentialHumanFlow-package.R
\docType{package}
\name{HodgePotentialHumanFlow-package}
\alias{HodgePotentialHumanFlow}
\alias{HodgePotentialHumanFlow-package}
\title{Scalar potential of origin-destionation matrix by Hodge-Kodaira decompostion}
\description{
This package implements Hodge-Kodaira decomposition of origin-destination matrix
}
\author{
\strong{Maintainer}: Takaaki Aoki \email{takaaki.aoki.work@gmail.com} (\href{https://orcid.org/0000-0002-3312-3061}{ORCID})

}
\keyword{internal}
